backend/requirements.txt
Django
django-flower
django-redis
djangorestframework
dotenv
psycopg2-binary
gunicorn
whitenoise
celery
redis
flower>=2.0
